refactor(scripts): Merge deploy and update DAT scripts into onw single DAT.js file in scripts
